+++ Task-10_Browser-game/src/js/rpgSystem.js
@@ -100,17 +100,3 @@ export function getDamageByStrength(strength) {
     }
     return result;
 }
-
-export function getDefaultCharacter() {
-    let character = {
-        states: getStatesTemplate(),
-        armor: {
-            [ArmorParts.head]: 0,
-            [ArmorParts.body]: 0,
-            [ArmorParts.hands]: 0,
-            [ArmorParts.legs]: 0,
-        },
-        weapon: 0
-    };
-    return character;
-}
